install: follow config_path symlink
Under NixOS, the config_path /etc/systemd/system is a symlink to
/etc/static/systemd/system. Commands such as `systemctl list-unit-files`
and `systemctl is-enabled` did not work as the symlink was not followed.

As a rule systemd does not follow symlinks when reading install
information under /etc as that is reserved for configuration. It's
unclear whether that could ever be a problem in this case, for NixOS.

I am testing this locally now. Hopefully, this fixes
https://github.com/NixOS/nixpkgs/issues/1083